ISSUE-02: moved receipt.py to package pythonsupermarket. adapted dependent files that use it.
